requirements.txt
deepagents 
tavily-python 
httpx 
markdownify 
langchain-anthropic 
langchain-core
python-dotenv